Refactor to remove other template features.
