+++ src/app/page.tsx
@@ -1,137 +1,72 @@
-import { ApiKeyCheck } from "@/components/ApiKeyCheck";
-import Image from "next/image";
+"use client";
 
-const KeyFilesSection = () => (
-  <div className="bg-white px-8 py-4">
-    <h2 className="text-xl font-semibold mb-4">How it works:</h2>
-    <ul className="space-y-4 text-gray-600">
-      <li className="flex items-start gap-2">
-        <span>📄</span>
-        <span>
-          <code className="font-medium">src/app/layout.tsx</code> - Main layout
-          with TamboProvider
-        </span>
-      </li>
-      <li className="flex items-start gap-2">
-        <span>📄</span>
-        <span>
-          <code className="font-medium font-mono">src/app/chat/page.tsx</code> -
-          Chat page with TamboProvider and MCP integration
-        </span>
-      </li>
-      <li className="flex items-start gap-2">
-        <span>📄</span>
-        <span>
-          <code className="font-medium font-mono">
-            src/app/interactables/page.tsx
-          </code>{" "}
-          - Interactive demo page with tools and components
-        </span>
-      </li>
-      <li className="flex items-start gap-2">
-        <span>📄</span>
-        <span>
-          <code className="font-medium font-mono">
-            src/components/tambo/message-thread-full.tsx
-          </code>{" "}
-          - Chat UI
-        </span>
-      </li>
-      <li className="flex items-start gap-2">
-        <span>📄</span>
-        <span>
-          <code className="font-medium font-mono">
-            src/components/tambo/graph.tsx
-          </code>{" "}
-          - A generative graph component
-        </span>
-      </li>
-      <li className="flex items-start gap-2">
-        <span>📄</span>
-        <span>
-          <code className="font-medium font-mono">
-            src/services/population-stats.ts
-          </code>{" "}
-          - Example tool implementation with mock population data
-        </span>
-      </li>
-      <li className="flex items-start gap-2">
-        <span className="text-blue-500">📄</span>
-        <span>
-          <code className="font-medium font-mono">src/lib/tambo.ts</code> -
-          Component and tool registration
-        </span>
-      </li>
-      <li className="flex items-start gap-2">
-        <span className="text-blue-500">📄</span>
-        <span>
-          <code className="font-medium font-mono">README.md</code> - For more
-          details check out the README
-        </span>
-      </li>
-    </ul>
-    <div className="flex gap-4 flex-wrap mt-4">
-      <a
-        href="https://docs.tambo.co"
-        target="_blank"
-        rel="noopener noreferrer"
-        className="px-6 py-3 rounded-md font-medium transition-colors text-lg mt-4 border border-gray-300 hover:bg-gray-50"
-      >
-        View Docs
-      </a>
-      <a
-        href="https://tambo.co/dashboard"
-        target="_blank"
-        rel="noopener noreferrer"
-        className="px-6 py-3 rounded-md font-medium transition-colors text-lg mt-4 border border-gray-300 hover:bg-gray-50"
-      >
-        Dashboard
-      </a>
-    </div>
-  </div>
-);
+import { useMcpServers } from "@/components/tambo/mcp-config-modal";
+import { MessageThreadFull } from "@/components/tambo/message-thread-full";
+import { SimpleMusicPlayer } from "@/components/ui/SimpleMusicPlayer";
+import { components, tools } from "@/lib/tambo";
+import { TamboProvider } from "@tambo-ai/react";
+import { TamboMcpProvider } from "@tambo-ai/react/mcp";
 
 export default function Home() {
+  const mcpServers = useMcpServers();
+
   return (
-    <div className="min-h-screen p-8 flex flex-col items-center justify-center font-[family-name:var(--font-geist-sans)]">
-      <main className="max-w-2xl w-full space-y-8">
-        <div className="flex flex-col items-center">
-          <a href="https://tambo.co" target="_blank" rel="noopener noreferrer">
-            <Image
-              src="/Octo-Icon.svg"
-              alt="Tambo AI Logo"
-              width={80}
-              height={80}
-              className="mb-4"
-            />
-          </a>
-          <h1 className="text-4xl text-center">tambo-ai chat template</h1>
-        </div>
+    <div className="h-screen flex flex-col overflow-hidden relative">
+      <TamboProvider
+        apiKey={process.env.NEXT_PUBLIC_TAMBO_API_KEY!}
+        components={components}
+        tools={tools}
+        tamboUrl={process.env.NEXT_PUBLIC_TAMBO_URL}
+      >
+        <TamboMcpProvider mcpServers={mcpServers}>
+          <div className="h-full w-full flex overflow-hidden">
+            <div className="w-[520px] min-w-[380px] max-w-lg h-full border-r overflow-auto">
+              <SimpleMusicPlayer 
+                tracks={[
+                  {
+                    title: "SoundHelix Song 1",
+                    artist: "SoundHelix",
+                    album: "Demo Album",
+                    duration: 30,
+                    preview: "https://www.soundhelix.com/examples/mp3/SoundHelix-Song-1.mp3",
+                    link: "https://www.soundhelix.com/examples/mp3/SoundHelix-Song-1.mp3",
+                    albumCover: "https://picsum.photos/seed/1/256",
+                  },
+                  {
+                    title: "SoundHelix Song 2",
+                    artist: "SoundHelix",
+                    album: "Demo Album", 
+                    duration: 30,
+                    preview: "https://www.soundhelix.com/examples/mp3/SoundHelix-Song-2.mp3",
+                    link: "https://www.soundhelix.com/examples/mp3/SoundHelix-Song-2.mp3",
+                    albumCover: "https://picsum.photos/seed/2/256",
+                  },
+                  {
+                    title: "SoundHelix Song 3",
+                    artist: "SoundHelix", 
+                    album: "Demo Album",
+                    duration: 30,
+                    preview: "https://www.soundhelix.com/examples/mp3/SoundHelix-Song-3.mp3",
+                    link: "https://www.soundhelix.com/examples/mp3/SoundHelix-Song-3.mp3",
+                    albumCover: "https://picsum.photos/seed/3/256",
+                  },
+                ]}
+                currentIndex={0}
+                isPlaying={false}
+              />
+            </div>
 
-        <div className="w-full space-y-8">
-          <div className="bg-white px-8 py-4">
-            <h2 className="text-xl font-semibold mb-4">Setup Checklist</h2>
-            <ApiKeyCheck>
-              <div className="flex gap-4 flex-wrap">
-                <a
-                  href="/chat"
-                  className="px-6 py-3 rounded-md font-medium shadow-sm transition-colors text-lg mt-4 bg-[#7FFFC3] hover:bg-[#72e6b0] text-gray-800"
-                >
-                  Go to Chat →
-                </a>
-                <a
-                  href="/interactables"
-                  className="px-6 py-3 rounded-md font-medium shadow-sm transition-colors text-lg mt-4 bg-[#FFE17F] hover:bg-[#f5d570] text-gray-800"
-                >
-                  Interactables Demo →
-                </a>
+            <div className="flex-1 min-w-0 h-full">
+              <div className="h-full flex flex-col">
+                <MessageThreadFull
+                  className="right w-full max-w-none ml-0"
+                  contextKey="tambo-template"
+                />
               </div>
-            </ApiKeyCheck>
+            </div>
           </div>
-
-          <KeyFilesSection />
-        </div>
-      </main>
+        </TamboMcpProvider>
+      </TamboProvider>
     </div>
   );
 }
