+++ src/services/music-search.ts
@@ -9,7 +9,7 @@ export async function searchMusic({ query }: SearchMusicInput): Promise<Song[]> 
   const data = await response.json();
   if (data.error) throw new Error(data.error);
   // Map Deezer API results to MusicSearchResult type
-  return (data.data || []).map((item: any) => ({
+  return (data.data || []).map((item: { artist: { name: string }; album: { title: string; cover: string }; [key: string]: unknown }) => ({
     ...item,
     artist: item.artist?.name || "",
     album: item.album?.title || "",
